requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
asyncgui==0.1.3
asynctkinter==0.1.1
attrs==20.3.0
audioread==2.1.9
azure-cognitiveservices-speech==1.16.0
backcall==0.2.0
bleach==3.3.0
bokeh==2.3.1
boto3==1.17.99
botocore==1.20.99
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorcet==2.0.6
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.21
decorator
defusedxml==0.7.1
entrypoints==0.3
gensim==4.0.1
holoviews==1.14.3
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
mutagen==1.45.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
noisereduce==1.1.0
notebook==6.3.0
numba==0.53.1
numpy==1.20.1
olefile==0.46
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
panel==0.11.3
param==1.10.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
pooch==1.3.0
proglog==0.1.9
prometheus-client==0.10.1
prompt-toolkit==3.0.17
pycparser==2.20
pyct==0.4.8
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-qt==0.50
python-vlc==3.0.11115
pytz==2021.1
pyviz-comms==2.0.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
regex==2021.3.17
requests==2.25.1
resampy==0.2.2
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
sip==6.1.1
six==1.15.0
smart-open==5.0.0
SoundFile==0.10.3.post1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.4.1

PyMySQL==1.0.2
fpdf==1.7.2
google-api-python-client==2.10.0
wordcloud==1.8.1
opencv-python==4.5.2.52